Consolidate common code between HttpResponseStream managed and Windows
